# nodejs-server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "CodeNaive Tools API"
  version: "1.0.0"
  title: "CodeNaive API"
  contact:
    email: "S.8506.A@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/codenaive/Hello/1.0.0"
tags:
- name: "admins"
  description: "Secured Admin-only calls"
- name: "developers"
  description: "Operations available to regular developers"
schemes:
- "https"
paths:
  /codenaive_data:
    get:
      tags:
      - "developers"
      summary: "searches data"
      description: "By passing in the appropriate options, you can search for\navailable\
        \ inventory in the system\n"
      operationId: "searchData"
      produces:
      - "application/json"
      parameters:
      - name: "searchString"
        in: "query"
        description: "pass an optional search string for looking up inventory"
        required: false
        type: "string"
      - name: "skip"
        in: "query"
        description: "number of records to skip for pagination"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "maximum number of records to return"
        required: false
        type: "integer"
        maximum: 50
        minimum: 0
        format: "int32"
      responses:
        "200":
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/codenaive_dataItem"
        "400":
          description: "bad input parameter"
      x-swagger-router-controller: "Developers"
    post:
      tags:
      - "admins"
      summary: "adds an data item"
      description: "Adds an item to the system"
      operationId: "addData"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "codenaive_dataItem"
        description: "Data item to add"
        required: false
        schema:
          $ref: "#/definitions/codenaive_dataItem"
      responses:
        "201":
          description: "item created"
        "400":
          description: "invalid input, object invalid"
        "409":
          description: "an existing item already exists"
      x-swagger-router-controller: "Admins"
definitions:
  codenaive_dataItem:
    type: "object"
    required:
    - "id"
    - "manufacturer"
    - "name"
    - "releaseDate"
    properties:
      id:
        type: "string"
        format: "uuid"
        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      name:
        type: "string"
        example: "CodeNaive"
      releaseDate:
        type: "string"
        format: "date-time"
        example: "2021-01-01T09:00:00.000+0000"
      manufacturer:
        $ref: "#/definitions/Manufacturer"
    example:
      releaseDate: "2021-01-01T09:00:00.000+0000"
      name: "CodeNaive"
      id: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      manufacturer:
        phone: "408-867-5309"
        name: "CodeNaive Corporation"
        homePage: "https://codenaive.com"
  Manufacturer:
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "CodeNaive Corporation"
      homePage:
        type: "string"
        format: "url"
        example: "https://codenaive.com"
      phone:
        type: "string"
        example: "408-867-5309"
    example:
      phone: "408-867-5309"
      name: "CodeNaive Corporation"
      homePage: "https://codenaive.com"